manually bump version
The v0.0.9 release partially failed, travis published v0.0.9 to pypi
but failed to bump the version therefore we need to manually
bump the version so travis won't try to publish the same version
to pypi again on the next release. This sets travis ci up to do
the next publish and version bumnp.