Add Basic Config File System
Had to remove key-file as an option. Hypothetical situation I'm worried about is: there is a key specified in the DEFAULT section of a config, and a key-file specified in a profile. The CLI would skip over the `key-file` in the profile, even when that profile was explicitly selected, and jump to the `key` in DEFAULT. Have to think of a way around this.